+++ .script_sizes.json
@@ -1,7 +1,7 @@
 {
   "scripts/cluster/__init__.py": 0,
   "scripts/cluster/modeling.py": 29429,
-  "scripts/cluster/simulator.py": 17614,
+  "scripts/cluster/simulator.py": 24417,
   "scripts/cluster/start_here.py": 18571,
   "scripts/group/__init__.py": 0,
   "scripts/group/data_preparation/__init__.py": 0,
@@ -96,10 +96,9 @@
   "scripts/guides/results/aggregator/queries.py": 5125,
   "scripts/guides/results/aggregator/samples.py": 20634,
   "scripts/guides/results/aggregator/samples_via_aggregator.py": 21443,
-  "scripts/guides/results/aggregator/start_here.py": 19956,
   "scripts/guides/results/database/__init__.py": 0,
   "scripts/guides/results/database/start_here.py": 14466,
-  "scripts/guides/results/start_here.py": 5824,
+  "scripts/guides/results/start_here.py": 26719,
   "scripts/guides/results/workflow/__init__.py": 0,
   "scripts/guides/results/workflow/csv_make.py": 15706,
   "scripts/guides/results/workflow/fits_make.py": 16665,
@@ -115,11 +114,11 @@
   "scripts/imaging/data_preparation/examples/data.py": 8898,
   "scripts/imaging/data_preparation/examples/noise_map.py": 5587,
   "scripts/imaging/data_preparation/examples/optional/__init__.py": 0,
-  "scripts/imaging/data_preparation/examples/optional/extra_galaxies_centres.py": 4449,
+  "scripts/imaging/data_preparation/examples/optional/extra_galaxies_centres.py": 4489,
   "scripts/imaging/data_preparation/examples/optional/info.py": 3249,
   "scripts/imaging/data_preparation/examples/optional/lens_light_centre.py": 3284,
   "scripts/imaging/data_preparation/examples/optional/mask.py": 4980,
-  "scripts/imaging/data_preparation/examples/optional/mask_extra_galaxies.py": 4424,
+  "scripts/imaging/data_preparation/examples/optional/mask_extra_galaxies.py": 5213,
   "scripts/imaging/data_preparation/examples/optional/positions.py": 3283,
   "scripts/imaging/data_preparation/examples/psf.py": 6647,
   "scripts/imaging/data_preparation/gui/__init__.py": 0,
@@ -165,8 +164,8 @@
   "scripts/imaging/features/advanced/subhalo/sensitivity/start_here.py": 35179,
   "scripts/imaging/features/advanced/subhalo/simulator.py": 9178,
   "scripts/imaging/features/extra_galaxies/__init__.py": 0,
-  "scripts/imaging/features/extra_galaxies/modeling.py": 21075,
-  "scripts/imaging/features/extra_galaxies/simulator.py": 8968,
+  "scripts/imaging/features/extra_galaxies/modeling.py": 20791,
+  "scripts/imaging/features/extra_galaxies/simulator.py": 10667,
   "scripts/imaging/features/extra_galaxies/slam.py": 17116,
   "scripts/imaging/features/linear_light_profiles/__init__.py": 0,
   "scripts/imaging/features/linear_light_profiles/fit.py": 8986,
@@ -182,15 +181,15 @@
   "scripts/imaging/features/multi_gaussian_expansion/source_science.py": 12754,
   "scripts/imaging/features/no_lens_light/__init__.py": 0,
   "scripts/imaging/features/no_lens_light/modeling.py": 10038,
-  "scripts/imaging/features/no_lens_light/simulator.py": 6630,
+  "scripts/imaging/features/no_lens_light/simulator.py": 7532,
   "scripts/imaging/features/no_lens_light/slam.py": 11533,
   "scripts/imaging/features/pixelization/__init__.py": 0,
   "scripts/imaging/features/pixelization/adaptive.py": 19767,
   "scripts/imaging/features/pixelization/cpu_fast_modeling.py": 19013,
   "scripts/imaging/features/pixelization/delaunay.py": 52363,
-  "scripts/imaging/features/pixelization/fit.py": 24225,
+  "scripts/imaging/features/pixelization/fit.py": 23941,
   "scripts/imaging/features/pixelization/likelihood_function.py": 40615,
-  "scripts/imaging/features/pixelization/modeling.py": 25198,
+  "scripts/imaging/features/pixelization/modeling.py": 24914,
   "scripts/imaging/features/pixelization/slam.py": 14195,
   "scripts/imaging/features/pixelization/source_science.py": 18837,
   "scripts/imaging/features/scaling_relation/__init__.py": 0,
@@ -198,7 +197,7 @@
   "scripts/imaging/features/simulator_manual_signal_to_noise_ratio.py": 7606,
   "scripts/imaging/fit.py": 13456,
   "scripts/imaging/likelihood_function.py": 18952,
-  "scripts/imaging/modeling.py": 28503,
+  "scripts/imaging/modeling.py": 28443,
   "scripts/imaging/simulator.py": 11931,
   "scripts/imaging/simulator_sample.py": 8636,
   "scripts/imaging/source_science.py": 10244,
@@ -210,7 +209,7 @@
   "scripts/interferometer/data_preparation/examples/__init__.py": 0,
   "scripts/interferometer/features/__init__.py": 0,
   "scripts/interferometer/features/extra_galaxies/__init__.py": 0,
-  "scripts/interferometer/features/extra_galaxies/modeling.py": 12736,
+  "scripts/interferometer/features/extra_galaxies/modeling.py": 12743,
   "scripts/interferometer/features/extra_galaxies/simulator.py": 8209,
   "scripts/interferometer/features/extra_galaxies/slam.py": 16144,
   "scripts/interferometer/features/pixelization/__init__.py": 0,
